+++ R/get_parameters.R
@@ -11,6 +11,7 @@
 #'      \item{\link[=get_parameters.glmm]{Mixed models} (\pkg{lme4}, \pkg{glmmTMB}, \pkg{GLMMadaptive}, ...)}
 #'      \item{\link[=get_parameters.zeroinfl]{Zero-inflated and hurdle models} (\pkg{pscl}, ...)}
 #'      \item{\link[=get_parameters.betareg]{Models with special components} (\pkg{betareg}, \pkg{MuMIn}, ...)}
+#'      \item{\link[=get_parameters.htest]{Hypothesis tests} (\code{htest})}
 #'    }
 #'
 #' @param verbose Toggle messages and warnings.